requirements.txt
google-auth==2.33.0
gspread==6.1.2
pandas==2.2.2
python-dotenv==1.0.1
email-validator==2.2.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.113.0
openai==1.51.2
httpx==0.27.2
anthropic>=0.31.0